% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_sql_dir}
\alias{create_sql_dir}
\title{Generate Helper Functions}
\usage{
create_sql_dir(context)
}
\arguments{
\item{context}{The context of your project.  "project", "shiny", and "sandbox"}
}
\description{
\code{create_sql_dir} will generate the path used to house the sql files based on the context parameter.
}
