% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_sql.R
\name{display_sql_with_comments}
\alias{display_sql_with_comments}
\title{Display sql with comments}
\usage{
display_sql_with_comments(folder, file)
}
\arguments{
\item{folder}{The name of the folder from inst containing the sql file.}

\item{file}{The name of the sql file you want to display.}
}
\value{
A sql file in your SQL folder
}
\description{
Display sql with comments
}
