% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fte_sql.R
\name{make_fte_sql}
\alias{make_fte_sql}
\title{\code{make_fte_sql} function will generate the standard SQL query for pulling Full-Time
Equivalent (FTE) data from Edify.This function allows users to retrieve FTE data based on
different available data sources. Use the \code{type} parameter to select the specific FTE data
type. Available options are: "current", "census" and "eot" (end of term). For more details see
the fte (full time quivalent) vignette. You can bring this up by running the following code in
your console: \code{vignette("full_time_equivalent", package = "utDataStoR")}
#' \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
make_fte_sql(name, type = "current")
}
\arguments{
\item{name}{The name you want the SQL file to have in your sql folder. This is a string and
must be provided in quotes (e.g., "your_filename.sql").}

\item{type}{The type of fte file you want. Defaults to 'current' fte data.
Other options are
\itemize{
\item 'current' -- data as of the current snapshot
\item 'census_demographic' -- fte data with demographic data, as of census snapshot
\item 'eot' -- data as of the end of term snapshot
}}
}
\value{
A sql file in your SQL folder
}
\description{
\code{make_fte_sql} function will generate the standard SQL query for pulling Full-Time
Equivalent (FTE) data from Edify.This function allows users to retrieve FTE data based on
different available data sources. Use the \code{type} parameter to select the specific FTE data
type. Available options are: "current", "census" and "eot" (end of term). For more details see
the fte (full time quivalent) vignette. You can bring this up by running the following code in
your console: \code{vignette("full_time_equivalent", package = "utDataStoR")}
#' \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
