% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_retention_sql.R
\name{make_retention_sql}
\alias{make_retention_sql}
\title{Generate Retention Sql}
\usage{
make_retention_sql(name, type = "term_to_term")
}
\arguments{
\item{name}{The name you want the SQL file to have in your sql folder. This is a string and must be provided in quotes
(e.g., "your_filename.sql").}

\item{type}{The type of retention file you want. Defaults to 'term_to_term'.

The type parameter should be
term_to_term
cohort}
}
\value{
A sql file in your SQL folder
}
\description{
\code{make_retention_sql} will generate the standard SQL query for pulling retention
data from Edify. This function helps in retrieving data related to student retention metrics.
To specify the type of retention data you would like, use the \code{type} parameter.
}
