% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fte.R
\name{sql_fte}
\alias{sql_fte}
\title{Generate FTE SQL
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
sql_fte(data_source = "Edify", name)
}
\arguments{
\item{data_source}{The server to pull FTE data from. Options are: "Edify", "REPT", or "DSCIR". Defaults to "Edify".}

\item{name}{The name you want the SQL file to have in your sql folder. This is a string and must be provided in quotes
(e.g., "your_filename.sql").}
}
\value{
A sql file in your SQL folder
}
\description{
\code{sql_fte} was intended to generate the standard sql query for pulling FTE from Edify, REPT
of DSCIR. We have deprecated this function and will use "make_fte_sql()" instead.
}
