% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utDataStoR-package.R
\docType{package}
\name{utDataStoR-package}
\alias{utDataStoR}
\alias{utDataStoR-package}
\title{utDataStoR: SQL and data sets for the Office of Institutional Effectiveness}
\description{
This package contains common SQL queries and data sets used by the Office of Institutional Effectiveness at Utah Tech University.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ut-effectiveness/utDataStoR}
  \item \url{https://ut-effectiveness.github.io/utDataStoR/}
  \item Report bugs at \url{https://github.com/ut-effectiveness/utDataStoR/issues}
}

}
\author{
\strong{Maintainer}: Craig Demke \email{craig.demke@utahtech.edu}

Authors:
\itemize{
  \item Matt Nickodemus \email{matt.nickodemus@utahtech.edu}
  \item Joy Baule \email{joy.baule@utahtech.edu}
  \item Andrea Bringhurst \email{andrea.bringhurst@utahtech.edu}
}

}
\keyword{internal}
