% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_sql.R
\name{display_sql_with_comments}
\alias{display_sql_with_comments}
\title{Display sql with comments}
\usage{
display_sql_with_comments(folder, file)
}
\arguments{
\item{folder}{A character string specifying the folder within \code{inst/sql} where the SQL file is located.
Must be within quotes.}

\item{file}{A character string specifying the name of the SQL file to display. Must be within quotes.}
}
\value{
A character string containing the content of the SQL file
}
\description{
This function returns the content of a specified SQL file, including any comments within
the file. The function should used within a sql code chunk (\code{sql } ) in R markdown within single back tick.
Please refer to the example section for better understanding.
}
