% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_headcount_sql.R
\name{make_headcount_sql}
\alias{make_headcount_sql}
\title{Generate headcount Sql}
\usage{
make_headcount_sql(name, type = "current")
}
\arguments{
\item{name}{The name you want the SQL file to have in your sql folder.}

\item{type}{The type of head count file you want. Defaults to 'current' headcount data.
Other options:
'census' data collected at 3rd week,
'eot' end of term data,
'ipeds' data sent to IPEDS,
and 'pit' point in time data,}
}
\value{
A sql file in your SQL folder
}
\description{
\code{make_headcount_sql} will generate the standard sql query for pulling headcount
data from Edify
}
