% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_retention_sql.R
\name{make_retention_sql}
\alias{make_retention_sql}
\title{Generate Retention Sql}
\usage{
make_retention_sql(name, type = "term_to_term")
}
\arguments{
\item{name}{The name you want the SQL file to have in your sql folder.}

\item{type}{The type of retention file you want. Defaults to 'term_to_term'.

The type parameter should be
term_to_term
cohort}
}
\value{
A sql file in your SQL folder
}
\description{
\code{make_retention_sql} will generate the standard sql query for pulling retention
data from Edify
}
