% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fte.R
\name{sql_fte}
\alias{sql_fte}
\title{Generate FTE SQL}
\usage{
sql_fte(data_source, name)
}
\arguments{
\item{data_source}{Bob. The server you will be full from. Edify, REPT, DSCIR}

\item{name}{The name you want the SQL file to have in your sql folder.}
}
\value{
A sql file in your SQL folder
}
\description{
\code{sql_fte} will generate the standard sql query for pulling FTE from Edify, REPT
of DSCIR.
}
