% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{sql_validate_student}
\alias{sql_validate_student}
\title{Generate Validation SQL}
\usage{
sql_validate_student(data_source, name, context)
}
\arguments{
\item{data_source}{The server you will be full from. Currently, Edify.}

\item{name}{The name you want the SQL file to have in your sql folder. This is a string and must be provided in quotes
(e.g., "your_filename.sql").}

\item{context}{The context of your project.  The options are: "project", "shiny", and "sandbox". The default is "project".}
}
\value{
A sql file in your SQL folder
}
\description{
\code{sql_validate_student} will generate the standard sql query for pulling Student Validations from Edify.
Validations are done through Edify.
}
